requirements.txt
Flask==1.0.2
Pillow==5.3.0
requests==2.21.0
scipy==1.1.0
